[IMP] {test_}mail: improve activity handling and add reuse support
This commit makes the following improvements for the activities.

- Remove the reschedule button from the done activities.
- Add a rainbow_man effect for empty systray opening.
- Add support to reuse the done activity from the posted message record.
- Add the mail_activity_id field on the message, as earlier the activity
  and message were never connected so this feature will not work for the older
  messages.

Task-5344015

+++ addons/mail/static/src/core/web/activity_model_patch.js
@@ -75,7 +75,15 @@ patch(Activity.prototype, {
         return action;
     },
     remove({ broadcast = true } = {}) {
-        this.delete();
+        // Only delete the activity if no message is linked to it.
+        // We keep it when marked as done because reuse action needs mail_activity_id.
+        if (
+            !Object.values(this.store["mail.message"].records).some(
+                (m) => m.mail_activity_id?.id === this.id
+            )
+        ) {
+            this.delete();
+        }
         if (broadcast) {
             this.activityBroadcastChannel?.postMessage({
                 type: "DELETE",
